pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hexpay"
version = "2.0.0"
description = "Official Python SDK for the HexPay Merchant API (TON cryptocurrency payments)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
authors = [{ name = "HexPay" }]
keywords = ["hexpay", "ton", "crypto", "cryptocurrency", "payments", "sdk", "api"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Typing :: Typed",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx >=0.23.0,<0.29.0",
    "attrs >=22.2.0",
    "python-dateutil >=2.8.0",
]

[project.urls]
Homepage      = "https://hexpay.io"
Documentation = "https://docs.hexpay.io"
Repository    = "https://github.com/hexpay/hexpay-python-sdk"
Issues        = "https://github.com/hexpay/hexpay-python-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["hexpay"]

[tool.hatch.build.targets.sdist]
include = [
    "hexpay/**",
    "README.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"hexpay/**" = ["E", "F", "I", "UP", "B"]  # generated code — do not lint
